+++ src/app/ui/annotation-controls.tsx
@@ -1,17 +1,16 @@
-import { IconButton, Button } from 'Molstar/mol-plugin-ui/controls/common';
-import { PurePluginUIComponent } from 'Molstar/mol-plugin-ui/base';
-import { StructureQualityReportColorThemeProvider } from 'Molstar/extensions/pdbe/structure-quality-report/color';
-import { StateSelection, StateTransform } from 'Molstar/mol-state';
-import { ParameterControls } from 'Molstar/mol-plugin-ui/controls/parameters';
-import { DomainAnnotationsColorThemeProvider } from '../domain-annotations/color';
 import { PDBeStructureQualityReport } from 'Molstar/extensions/pdbe/structure-quality-report/behavior';
-import { PDBeDomainAnnotations } from '../domain-annotations/behavior';
-import { Icon, ArrowRightSvg, ArrowDropDownSvg, VisibilityOffOutlinedSvg, VisibilityOutlinedSvg, MoreHorizSvg } from 'Molstar/mol-plugin-ui/controls/icons';
+import { StructureQualityReportColorThemeProvider } from 'Molstar/extensions/pdbe/structure-quality-report/color';
 import { StructureHierarchyManager } from 'Molstar/mol-plugin-state/manager/structure/hierarchy';
+import { PurePluginUIComponent } from 'Molstar/mol-plugin-ui/base';
+import { Button } from 'Molstar/mol-plugin-ui/controls/common';
+import { ArrowDropDownSvg, ArrowRightSvg, Icon } from 'Molstar/mol-plugin-ui/controls/icons';
+import { StateSelection, StateTransform } from 'Molstar/mol-state';
+import { PDBeDomainAnnotations } from '../domain-annotations/behavior';
+import { DomainAnnotationsColorThemeProvider } from '../domain-annotations/color';
 import { InitParams } from '../spec';
-import { ParamDefinition as PD } from 'molstar/lib/mol-util/param-definition';
-import { AssemblySymmetryProvider } from 'molstar/lib/extensions/rcsb/assembly-symmetry/prop';
 import { AnnotationRowControls } from './annotation-row-controls';
+import { SymmetryAnnotationControls } from './symmetry-annotation-controls';
+
 
 const _TextsmsOutlined = <svg width='24px' height='24px' viewBox='0 0 24 24'><path fill="none" d="M0 0h24v24H0V0z" /><g><path d="M20 2H4c-1.1 0-2 .9-2 2v18l4-4h14c1.1 0 2-.9 2-2V4c0-1.1-.9-2-2-2zm0 14H6l-2 2V4h16v12z" /><path d="M7 9h2v2H7zM11 9h2v2h-2zM15 9h2v2h-2z" /></g></svg>;
 export function TextsmsOutlinedSvg() { return _TextsmsOutlined; }
@@ -26,16 +25,11 @@ interface AnnotationsComponentControlsState {
     domainsApplied: boolean,
     domainsOptions: boolean,
     domainsParams: any,
-    symmetryApplied: boolean,
-    symmetryOptions: boolean,
-    symmetryParams: any,
     description?: string,
 }
-interface SymmetryParams { } // TODO fill
 
 
 export class AnnotationsComponentControls extends PurePluginUIComponent<{}, AnnotationsComponentControlsState> {
-
     state: AnnotationsComponentControlsState = {
         isCollapsed: false,
         validationApplied: false,
@@ -44,9 +38,6 @@ export class AnnotationsComponentControls extends PurePluginUIComponent<{}, Anno
         domainsApplied: false,
         domainsOptions: false,
         domainsParams: undefined,
-        symmetryApplied: false,
-        symmetryOptions: false,
-        symmetryParams: undefined,
     };
 
     componentDidMount() {
@@ -63,11 +54,8 @@ export class AnnotationsComponentControls extends PurePluginUIComponent<{}, Anno
         const initParams: InitParams | undefined = (this.plugin.customState as any)?.initParams; // TODO define getters/setters for manipulating custom state?
         const validationAnnotationCtrl = !!initParams?.validationAnnotation;
         const domainAnnotationCtrl = !!initParams?.domainAnnotation;
-        const symmetryAnnotationCtrl = !!initParams?.symmetryAnnotation;
 
-        if ((validationAnnotationCtrl && !this.state.validationParams)
-            || (domainAnnotationCtrl && !this.state.domainsParams)
-            || (symmetryAnnotationCtrl && !this.state.symmetryParams)) {
+        if ((validationAnnotationCtrl && !this.state.validationParams) || (domainAnnotationCtrl && !this.state.domainsParams)) {
 
             const structure = this.getStructure()?.data;
             if (structure) {
@@ -97,43 +85,21 @@ export class AnnotationsComponentControls extends PurePluginUIComponent<{}, Anno
                     }
                 }
 
-                if (symmetryAnnotationCtrl && !this.state.symmetryParams) {
-                    const symmetryActionsParams = PD.clone(structure ? AssemblySymmetryProvider.getParams(structure) : AssemblySymmetryProvider.defaultParams);
-                    if ((symmetryActionsParams as any).serverType) { // TODO remove this condition and "as any" once we have newer Mol* version
-                        (symmetryActionsParams as any).serverType.isHidden = true;
-                    }
-                    symmetryActionsParams.serverUrl.isHidden = true;
-                    PD.getDefaultValues(symmetryActionsParams);
-                    console.log('params', symmetryActionsParams)
-                    this.setState({
-                        symmetryParams: {
-                            params: symmetryActionsParams,
-                            values: PD.getDefaultValues(symmetryActionsParams),
-                        }
-                    });
-                    // TODO symmetry
-                }
             }
         }
     }
     getStructure = () => {
         const groupRef = StateSelection.findTagInSubtree(this.plugin.state.data.tree, StateTransform.RootRef, 'structure-component-static-polymer');
-        return groupRef ? this.plugin.state.data.select(groupRef)[0].obj : undefined;
+        return groupRef ? this.plugin.state.data.select(groupRef)[0]?.obj : undefined;
     }
 
     toggleCollapsed = () => {
         this.setState({ isCollapsed: !this.state.isCollapsed });
     }
 
-    toggleOptions = (type: AnnotationType) => {
-        if (type === 'validation') this.setState({ validationOptions: !this.state.validationOptions });
-        if (type === 'domains') this.setState({ domainsOptions: !this.state.domainsOptions });
-        if (type === 'symmetry') this.setState({ symmetryOptions: !this.state.symmetryOptions });
-    }
-
     applyAnnotation = (type: 'validation' | 'domains', visibleState: boolean, params?: any) => {
         // Defaults
-        let themeName: any = 'polymer-id';
+        let themeName: any = 'polymer-id';  // TODO set to real default, this is not default (see 1e94 assembly 2)
         let themePropsToAdd = PDBeStructureQualityReport;
         let themePropsToRemove = this.state.domainsParams ? PDBeDomainAnnotations : void 0;
 
@@ -173,14 +139,9 @@ export class AnnotationsComponentControls extends PurePluginUIComponent<{}, Anno
         }
     }
 
-    applySymmetry(visibleState: boolean, params: SymmetryParams) {
-        this.setState({ symmetryApplied: visibleState });
-    }
-
     toggleAnnotation = (type: AnnotationType) => {
         if (type === 'validation') this.applyAnnotation('validation', !this.state.validationApplied, this.state.validationParams.values);
         if (type === 'domains') this.applyAnnotation('domains', !this.state.domainsApplied, this.state.domainsParams.values);
-        if (type === 'symmetry') this.applySymmetry(!this.state.symmetryApplied, this.state.symmetryParams.values);
     };
 
     updateValidationParams = (val: any) => {
@@ -195,15 +156,8 @@ export class AnnotationsComponentControls extends PurePluginUIComponent<{}, Anno
         this.setState({ domainsParams: updatedParams });
         if (this.state.domainsApplied) this.applyAnnotation('domains', this.state.domainsApplied, val);
     }
-    updateSymmetryParams = (val: any) => {
-        const updatedParams = { ...this.state.symmetryParams };
-        updatedParams.values = val;
-        this.setState({ symmetryParams: updatedParams });
-        if (this.state.symmetryApplied) this.applySymmetry(this.state.symmetryApplied, val);
-    }
 
     render() {
-        console.log('state:', this.state)
         if (!this.state.validationParams && !this.state.domainsParams) return <></>;
 
         const brand = {
@@ -233,75 +187,8 @@ export class AnnotationsComponentControls extends PurePluginUIComponent<{}, Anno
                 <AnnotationRowControls title='Domain Annotations' shortTitle='Domains'
                     params={this.state.domainsParams?.params} values={this.state.domainsParams?.values} onChangeValues={this.updateDomainParams}
                     applied={this.state.domainsApplied} onChangeApplied={() => this.toggleAnnotation('domains')} />
-                <AnnotationRowControls title='Assembly Symmetry'
-                    params={this.state.symmetryParams?.params} values={this.state.symmetryParams?.values} onChangeValues={this.updateSymmetryParams} />
+                <SymmetryAnnotationControls />
             </>}
-
-            {/* <div style={{ height: 10 }}></div>
-            {!this.state.isCollapsed && this.state.validationParams &&
-                <div className='msp-flex-row'>
-                    <Button noOverflow className='msp-control-button-label' title={`Validation Report Annotations.`} style={{ textAlign: 'left' }}>
-                        Validation Report
-                    </Button>
-                    <IconButton onClick={() => this.toggleAnnotation('validation')} toggleState={false} svg={!this.state.validationApplied ? VisibilityOffOutlinedSvg : VisibilityOutlinedSvg} title={`Click to ${this.state.validationApplied ? 'Hide' : 'Show'} Validation Report Annotation`} small className='msp-form-control' flex />
-                    <IconButton onClick={() => this.toggleOptions('validation')} svg={MoreHorizSvg} title='Actions' toggleState={this.state.validationOptions} className='msp-form-control' flex />
-                </div>
-            }
-            {!this.state.isCollapsed && this.state.validationParams && this.state.validationOptions &&
-                <div style={{ marginBottom: '6px' }}>
-                    <div className="msp-accent-offset">
-                        <div className='msp-representation-entry'>
-                            <ParameterControls params={this.state.validationParams.params} values={this.state.validationParams.values} onChangeValues={this.updateValidationParams} />
-                        </div>
-                    </div>
-                </div>
-            }
-
-            {!this.state.isCollapsed && this.state.domainsParams &&
-                <div className='msp-flex-row'>
-                    <Button noOverflow className='msp-control-button-label' title={`Domain Annotations.`} style={{ textAlign: 'left' }}>
-                        Domains
-                    </Button>
-                    <IconButton onClick={() => this.toggleAnnotation('domains')} toggleState={false} svg={!this.state.domainsApplied ? VisibilityOffOutlinedSvg : VisibilityOutlinedSvg} title={`Click to ${this.state.domainsApplied ? 'Hide' : 'Show'} Domain Annotation`} small className='msp-form-control' flex />
-                    <IconButton onClick={() => this.toggleOptions('domains')} svg={MoreHorizSvg} title='Actions' toggleState={this.state.domainsOptions} className='msp-form-control' flex />
-
-                </div>
-            }
-            {!this.state.isCollapsed && this.state.domainsParams && this.state.domainsOptions &&
-                <div style={{ marginBottom: '6px' }}>
-                    <div className="msp-accent-offset">
-                        <div className='msp-representation-entry'>
-                            <ParameterControls params={this.state.domainsParams.params} values={this.state.domainsParams.values} onChangeValues={this.updateDomainParams} />
-                        </div>
-                    </div>
-                </div>
-            }
-
-            {!this.state.isCollapsed && this.state.symmetryParams &&
-                <div className='msp-flex-row'>
-                    <Button noOverflow className='msp-control-button-label' title={`Assembly Symmetry.`} style={{ textAlign: 'left' }}>
-                        Assembly Symmetry
-                    </Button>
-                    <IconButton onClick={() => this.toggleAnnotation('symmetry')} toggleState={false} svg={!this.state.symmetryApplied ? VisibilityOffOutlinedSvg : VisibilityOutlinedSvg} title={`Click to ${this.state.symmetryApplied ? 'Hide' : 'Show'} Assembly Symmetry`} small className='msp-form-control' flex />
-                    <IconButton onClick={() => this.toggleOptions('symmetry')} svg={MoreHorizSvg} title='Actions' toggleState={this.state.symmetryOptions} className='msp-form-control' flex />
-
-                </div>
-            }
-            {!this.state.isCollapsed && this.state.symmetryParams && this.state.symmetryOptions &&
-                <div style={{ marginBottom: '6px' }}>
-                    <div className="msp-accent-offset">
-                        <div className='msp-representation-entry'>
-                            {this.state.symmetryApplied &&
-                                <ParameterControls params={this.state.symmetryParams.params} values={this.state.symmetryParams.values} onChangeValues={this.updateSymmetryParams} />
-                                || <div className='msp-row-text'>
-                                    <div title='You must enable Assembly Symmetry before setting parameters'>Off</div>
-                                </div>
-                            }
-                        </div>
-                    </div>
-                </div>
-
-            } */}
         </div>;
     }
 }